pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Ghouter <Ghouter2002@mail.ru>"]
readme = "README.md"
packages = [
    { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.1"
ruff = "^0.12.4"
dotenv = "^0.9.9"
gunicorn = "^23.0.0"
python-dotenv = "^1.1.1"
psycopg2-binary = "^2.9.10"
validators = "^0.35.0"
requests = "^2.32.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
